Use 'insert' instead of 'fill'; rename files and update spoken words.
Have the functions in the .py file return a string instead of
performing the actual insertion, which allows users to customize the
formatting as they see fit.
